+++ omibio/cli/kmer_cli/__init__.py
@@ -9,8 +9,10 @@ def kmer_group():
 
 def register_commands():
     from .kmer_count import count
+    from .kmer_total import total
 
     kmer_group.add_command(count)
+    kmer_group.add_command(total)
 
 
 register_commands()