+++ omibio/cli/kmer_cli/kmer_count.py
@@ -2,11 +2,11 @@ import click
 from omibio.cli.kmer_cli import kmer_group
 from omibio.io import read_fasta_iter
 from omibio.analysis import kmer
-from collections import Counter
 import sys
+import csv
 
 
-@kmer_group.command(name="kmer")
+@kmer_group.command()
 @click.argument(
     "fasta_file",
     type=click.File("r"),
@@ -25,7 +25,12 @@ import sys
 @click.option(
     "--top",
     type=int,
-    default=10
+)
+@click.option(
+    "--output", "-o",
+    type=click.Path(),
+    default=None,
+    help="Write details to a file in csv format"
 )
 @click.option(
     "--canonical", "-c",
@@ -36,29 +41,35 @@ def count(
     k: int,
     min_count: bool,
     canonical: bool,
-    top: int
+    output: str,
+    top: int | None,
 ):
     """Count k-mers in a FASTA file."""
     fh = fasta_file or sys.stdin
     entries = read_fasta_iter(fh)
 
-    total_counts: dict[str, int] = Counter()
-    total = 0
+    results: list[list[str | int]] = []
 
     for entry in entries:
         counts = kmer(
             entry.seq, k=k, canonical=canonical, min_count=min_count
         )
         for km, c in counts.items():
-            total_counts[km] += c
-            total += c
+            results.append([entry.seq_id, k, km, c])
 
     tops = sorted(
-        total_counts.keys(), key=lambda n: total_counts[n], reverse=True
+        results, reverse=True, key=lambda res: res[-1]
     )
 
-    click.echo(f"k = {k}")
-    click.echo(f"Total:\t{k}")
-    click.echo(f"Unique:\t{len(total_counts)}\n")
-    for top_kmer in tops[:top]:
-        click.echo(f"{top_kmer}:\t{total_counts[top_kmer]}")
+    if top is not None:
+        rows = [["seq_id", "k", "kmer", "count"]] + tops[:top]
+    else:
+        rows = [["seq_id", "k", "kmer", "count"]] + results
+    if output is not None:
+        with open(output, "w", newline="", encoding="utf-8") as f:
+            writer = csv.writer(f, delimiter="\t")
+            writer.writerows(rows)
+            click.echo(f"Written to {output}")
+    else:
+        for row in rows:
+            click.echo("\t".join(map(str, row)))