+++ omibio/viz/plot_kmer_heatmap.py
@@ -1,45 +1,59 @@
-import pandas as pd
-import seaborn as sns
-from matplotlib.axes import Axes
 import matplotlib.pyplot as plt
+import seaborn as sns
+import pandas as pd
 
 
 def plot_kmer(
-    kmer_counts: dict[str, int],
+    kmer_counts: list[dict[str, int]] | dict[str, int],
+    top_n: int = 20,
     ax=None,
-    cmap='Blues',
+    cmap="viridis",
     annot=True,
     fmt=".2f",
     show: bool = False
-) -> Axes:
+):
+    if isinstance(kmer_counts, dict):
+        kmer_counts = [kmer_counts]
+    kmers = []
+    k = len(list(kmer_counts[0].keys())[0])
+
+    for i, count in enumerate(kmer_counts):
+        kmers.append(pd.Series(count, name=f"seq_{i+1}"))
+    df = pd.DataFrame(kmers).fillna(0)
+
+    top_kmers = df.sum().sort_values(ascending=False).head(top_n).index
+    df_top = df[top_kmers]
+
+    df_top = df_top.div(df_top.sum(axis=1), axis=0)
+
     if ax is None:
-        ax = plt.subplots(figsize=(5, 5))[1]
-
-    df = pd.DataFrame.from_dict(kmer_counts, orient='index', columns=['count'])
-    df['freq'] = df['count'] / df['count'].sum()
-    plot_data = df[['freq']]
-
-    plot_data = plot_data.sort_values(by=plot_data.columns[0], ascending=False)
-
-    sns.heatmap(plot_data, ax=ax, cmap=cmap, annot=annot, fmt=fmt, cbar=True)
-    ax.set_ylabel('k-mer')
-    ax.set_xlabel('Sample')
-    ax.set_title('k-mer Frequency Heatmap')
-    ax.yaxis.set_tick_params(rotation=0)
+        ax = plt.subplots(
+            figsize=(max(8, top_n*0.5), max(6, len(kmer_counts)*0.5))
+        )[1]
+    sns.heatmap(df_top, annot=annot, fmt=fmt, cmap=cmap)
+    ax.set_ylabel("Sequence ID")
+    ax.set_xlabel(f"{k}-mers")
+    ax.set_title(f"{k}-mer Heatmap")
 
     if show:
         plt.show()
+
     return ax
 
 
 def main():
-    from omibio.analysis.kmer import kmer
-    from omibio.io import read_fasta
-    seq = read_fasta(
-        "./examples/data/example_single_short_seq.fasta"
-    )["example"]
-    kmer_counts = kmer(seq, k=2)
+    from omibio.io import read_fasta_iter
+    from omibio.analysis import kmer
+    kmer_counts = []
+    k = 3
+
+    source = "./examples/data/example_short_seqs.fasta"
+    for entry in read_fasta_iter(source):
+        kmer_counts.append(kmer(entry.seq, k))
+        if len(kmer_counts) == 5:
+            break
     plot_kmer(kmer_counts)
+    plt.tight_layout()
     plt.show()